- Changed Client::random_ping and Client::serial to standard unsigned integer   types
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid@653 82007160-df01-0410-b94d-b575c5fd34c7

+++ src/s_user.c
@@ -281,8 +281,8 @@ register_local_user(struct Client *source_p, const char *username)
   {
     if (!IsPingSent(source_p) && !source_p->localClient->random_ping)
     {
-      source_p->localClient->random_ping = (unsigned long)rand();
-      sendto_one(source_p, "PING :%lu",
+      source_p->localClient->random_ping = (unsigned int)rand();
+      sendto_one(source_p, "PING :%u",
                  source_p->localClient->random_ping);
       SetPingSent(source_p);
       return;
