+++ src/send.c
@@ -42,7 +42,7 @@ struct Callback *iosendctrl_cb = NULL;
 static void send_message(struct Client *, char *, int);
 static void send_message_remote(struct Client *, struct Client *, char *, int);
 
-static unsigned long current_serial = 0L;
+static unsigned int current_serial = 0;
 
 /* send_format()
  *
